% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.gui.R
\name{analysis.gui}
\alias{analysis.gui}
\title{Run the graphical user interface for analysis and computing the confidence interval
and concentration}
\usage{
analysis.gui()
}
\description{
Run the graphical user interface
}
\examples{
analysis.gui()
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
