% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apta.conf.R
\name{apta.conf}
\alias{apta.conf}
\title{To compute confidence interval of the combined replicates}
\usage{
apta.conf(aptakandt, ...)
}
\arguments{
\item{aptakandt}{an object of class \code{"aptakandt"}, which is an output of the function \code{CombineReps}.}
}
\value{
The object of the Class \code{aptakandt}, where the result will be saved in the slot confData.
}
\description{
Function \code{apta.conf} can be implemented on an object of class \code{"aptakandt"},
which is produced as an output from the function \code{CombineReps} and saves the results into a new
slot confData of class \code{"aptakandt"}. See \code{\link[AptaKan]{ReadAptakan}} and
\code{\link[AptaKan]{CombineReps}} also.
}
\details{
Function \code{apta.conf} computes the 95 percent confidence intervals of the combined replicates
which are produced as a result after implementing the function \code{CombineReps}. More details are
provided in the \code{AptaKan} package vignette.
}
\examples{
## Here are the examples are shown only with the .csv files and combining replicates on the basis of mean.
##Similarly, they can also be implemented on the other results of combining the replicates on the basis of
## median too through function CombineReps, For complete detail, see the vignettes
##To read the .csv file.
file1 <- system.file("exData", "Aptakandata.csv", package = "AptaKan")
read.datacsv <- ReadAptakan(file1)

## To combine the replicates on the basis of mean
cr.mean <- CombineReps(read.datacsv)
cr.mean ## To express all the data
slot(cr.mean,"crepData")  ##To express the combined replicates

## Compute confidence interval
conf <- apta.conf(cr.mean)
conf ##To visualise all the data
slot(conf,"confData") ## to express the confidence intervals
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{CombineReps,}
\keyword{ReadAptaKan,}
\keyword{aptakan.conf}
\keyword{aptakandt,}
