% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apta.Kdall.R
\docType{methods}
\name{apta.Kdall}
\alias{apta.Kdall}
\title{To compute the dissociation constant and plot the fitting results}
\usage{
\S4method{apta.Kdall}{aptakandt}(aptakandt, plot = "No", method = "sig",
  kdstart = 1, xlabel = "Concentration")
}
\arguments{
\item{aptakandt}{an object of class \code{"aptakandt"}, which is the output of function \code{CombineReps}.}

\item{plot}{To get the result in table form, mention "No". To get the plots mention "yes". Default is "No".}

\item{method}{To implement the sigmoidal model, mention "sig", while implementing the non-sigmoidal model
implement the "non-sig". The default is "sig".}

\item{xlabel}{To label the x-axis of the plot.}

\item{Kdstart}{The initial value to simulate the dissociation constant and it's curve. The default
value is 1.}
}
\value{
an object of Class \code{data.frame} or plots.
}
\description{
Function \code{apta.Kdall} can be implemented on an object of class \code{"aptakandt"},
which is produced as an output from the function \code{CombineReps} and saves the results into
class \code{"aptakandt"} and also the plots. The function \code{apta.Kdall} is based on the function
\code{apta.Kd}. See \code{\link[Aptakan]{ReadAptakan}}, \code{\link[AptaKan]{CombineReps}}
and \code{\link[AptaKan]{apta.Kd}} also.
}
\details{
Function \code{apta.Kdall} computes the dissociation constant in a single step of all
the combined replicates which are produced as a result after implementing the function
\code{CombineReps}. It also computes the correlation betweeen predicted and original data.
The function also plots the fitting curve of dissociation constant. The function \code{apta.kdall}
is based on the function \code{apta.Kd}. More details are provided in the \code{AptaKan} package
vignette.
}
\examples{
##To read the .csv file
file1 <- system.file("exData", "aptakdnew.csv", package = "AptaKan")
read.datacsv <- ReadAptakan(file1)

## To express all the data
read.datacsv

## To combine the replicates
kd.rep1.mean <- CombineReps(read.datacsv)
kd.rep1 ## To express all the data
slot(kd.rep1,"crepData")  ##To express the combined replicates

## To compute the dissociation constant by sigmoidal model
kd <- apta.Kdall(kd.rep1, kdstart = 1)

## To visualise the results of Kd and correlation between the original data and predicted
#data

slot(kd,"KdData")

## To plot the fitting results of the sigmoidal model
apta.Kdall(kd.rep1, plot = "yes", kdstart = 1)

## To compute the dissociation constant by non-sigmoidal model
kd.nonsig <- apta.Kdall(kd.rep1, method = "non-sig", kdstart = 1)

## To visualise the results of Kd and correlation between the original data and predicted
#data

slot(kd.nonsig,"KdData")

## To plot the fitting results of the sigmoidal model
apta.Kdall(kd.rep1, plot = "yes", method = "non-sig", kdstart = 1)
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{CombineReps,}
\keyword{ReadAptaKan,}
\keyword{apta.Kd,}
\keyword{apta.Kdall}
\keyword{aptakandt,}
