% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aptaconc.R
\docType{methods}
\name{aptaconc}
\alias{aptaconc}
\title{To compute the concentration from fluorescence}
\usage{
\S4method{aptaconc}{aptakandt}(aptakandt, fluo, model = "lm")
}
\arguments{
\item{aptakandt}{an object of class \code{"aptakandt"}, which is an output of the function \code{aptamodelall}.}

\item{fluo}{fluorescence value for which the concentration is to be calculated. It should be a numeric
value provided by the user.}

\item{model}{"lm" for linear model and robust linear model, while "loglm" for log linear model and robust log
linear model.}
}
\value{
an object of Class \code{data.frame}.
}
\description{
Function \code{aptaconc} can be implemented on an object of class \code{"aptakandt"},
which is produced as an output from the function \code{aptamodelall} and saves the results into
\code{data.frame}. See \code{\link[Aptakan]{ReadAptakan}}, \code{\link[AptaKan]{CombineReps}} and
\code{\link[AptaKan]{aptamodel}} and \code{\link[AptaKan]{aptamodelall}} also.
}
\details{
Function \code{aptaconc} computes the concentration of the combined replicates which are produced
as a result after implementing the function \code{aptamodelall}. One needs to provide fluorescence for which
the concentration is to be calculated.
}
\examples{
## Here the examples are shown only with the .csv files and combining replicates on the basis of mean.
##Similarly, they can also be implemented on the other results of combining the replicates on the basis of
## median too through function CombineReps, For complete detail, see the Vignettes.

##To read the .csv file.
file1 <- system.file("exData", "Aptakandata.csv", package = "AptaKan")
read.datacsv <- ReadAptakan(file1)

## To combine the replicates on the basis of mean
cr.mean <- CombineReps(read.datacsv)
cr.mean ## To express all the data
slot(cr.mean,"crepData")  ##To express the combined replicates

##To analyse the data on the basis of linear model
lm.result <- aptamodelall(cr.mean)
lm.result ## To express all the slots
slot(lm.result,"modelData") ##to express the modelData

## To analyse the data on the basis of log linear model
loglm.result <- aptamodelall(cr.mean, method = "loglm")
loglm.result
slot(loglm.result,"modelData")

## To analyse the data on the basis of robust linear model
roblm.result <- aptamodelall(cr.mean, method = "roblm")
roblm.result
slot(roblm.result,"modelData")

## To analyse the data on the basis of log robust linear model
logroblm.result <- aptamodelall(cr.mean, method = "logroblm")
logroblm.result
slot(logroblm.result,"modelData")

###Call the function aptaconc
apta.conc.lm <- aptaconc(lm.result,fluo = 1000)
apta.conc.lm
apta.conc.loglm <- aptaconc(loglm.result,fluo = 1019, model = "loglm")
apta.conc.loglm
apta.conc.roblm <- aptaconc(roblm.result,fluo = 1019, model = "lm")
apta.conc.roblm
apta.conc.logroblm <- aptaconc(logroblm.result,fluo = 1019, model = "loglm")
apta.conc.logroblm
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{CombineReps,}
\keyword{ReadAptaKan,}
\keyword{aptakan.conc}
\keyword{aptakandt,}
