% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{aptakandt-class}
\alias{aptakandt-class}
\title{S4 Class aptakandt to contain the data}
\section{Slots}{

\describe{
\item{\code{initialData}}{contains the initial data, except concentration, in the form of an object
of Class \code{data.frame}}

\item{\code{crepData}}{contains the result of the function \code{CombineReps} in the form of an object
of Class \code{data.frame}}

\item{\code{concentration}}{contains the concentration data in the form of an object of Class \code{data.frame}}

\item{\code{modelData}}{contains the result of implementation of different statistical models of
analysis in the form of an object of Class \code{data.frame}}

\item{\code{confData}}{contains the confidence intervals in the form of an object of Class \code{data.frame}}

\item{\code{KdData}}{contains the results of dissociation constant in the form of an object of Class
\code{data.frame}}
}}

