% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aptakanplot.R
\docType{methods}
\name{aptakanplot}
\alias{aptakanplot}
\title{To generate the plots on the basis of different models of lm, loglm, lmrob and loglmrob
to analyse the data.}
\usage{
\S4method{aptakanplot}{aptakandt}(aptakandt, method = "lm",
  main = "lm(x~y)")
}
\arguments{
\item{aptakandt}{an object of class \code{"aptakandt"}, which is an output of the function \code{CombineReps}.}

\item{method}{The default method is "lm", which is linearised model lm(y~x). The other options are
"loglm" for lm(log(y)~x),"roblm" for lmrob(y~x), which is robust linear model and "logroblm" for
lmrob(log(y)~x) to plot the results to analyse the data.}

\item{main}{The default is "lm(y~x)". The main heading of the plot can be defined by the user.
There is no need to define the main argument, While implementing through the function \code{aptakanplotall}.}
}
\value{
plots of the analysis of data
}
\description{
Function \code{aptakanplot} can be implemented on an object of class \code{"aptakandt"},
which is produced as an output from the function \code{CombineReps} and generates the plots on the
basis of different models of lm, loglm, lmrob and loglmrob. See \code{\link[Aptakan]{ReadAptakan}} and
\code{\link[AptaKan]{CombineReps}}. It's an auxilliary function to \code{aptakanplotall} function. It
can be implemented directly through the function \code{aptakanplotall}. See \code{\link[ggplot2]{qplot}}
also.
}
\details{
Allows the user to generate the plots of analysis of the data on the basis of different models,
including linear model,log linear model, robust linear model and log robust linear model. It provides
the different values of r-square, adj.rsquare, intercept and slope in case of linear, log linear, robust
linear and log robust linear models. This function \code{aptakanplot} is an auxilliary function to
the function \code{aptakanplotall}. This function \code{aptakanplot} needs not to be implemented alone.
It can implemented directly through the \code{aptakanplotall} function. More details are provided in
the \code{AptaKan} package vignette.
}
\examples{
## Here the examples are shown only with the .csv files and combining replicates on the basis of mean.
##Similarly, they can also be implemented on the other results of combining the replicates on the basis of
## median too through function CombineReps, For complete detail, see the vignettes.
##To read the .csv file.
file1 <- system.file("exData", "Aptakandata.csv", package = "AptaKan")
read.datacsv <- ReadAptakan(file1)

## To combine the replicates on the basis of mean
cr.mean <- CombineReps(read.datacsv)
cr.mean ## To express all the data
slot(cr.mean,"crepData")  ##To express the combined replicates

##plot on the basis of linear model
lm.plot <- aptakanplot(cr.mean)
lm.plot

##plot on the basis of log linear model
loglm.plot <- aptakanplot(cr.mean, method = "loglm", main = "lm(log(Y)~x)")
loglm.plot

##plot on the basis of robust linear model
roblm.plot <- aptakanplot(cr.mean, method = "roblm", main = "lmrob(y~x)")
roblm.plot

## plot on the basis of log robust linear model
logroblm.plot <- aptakanplot(cr.mean, method = "logroblm", main = "lmrob(log(y)~x)")
logroblm.plot
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{CombineReps,}
\keyword{ReadAptaKan,}
\keyword{aptakandt,}
\keyword{aptakanplot}
\keyword{aptamodelall,}
