% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aptamodel.R
\docType{methods}
\name{aptamodel}
\alias{aptamodel}
\title{To implement the different models of lm, loglm, lmrob, loglmrob and glm to analyse the data}
\usage{
\S4method{aptamodel}{aptakandt}(aptakandt, method = "lm", result = "lm")
}
\arguments{
\item{aptakandt}{an object of class \code{"aptakandt"}, which is an output of the function \code{CombineReps}.}

\item{method}{The default method is "lm", which is linearised model lm(y~x). The other options are
"loglm" for lm(log(y)~x),"roblm" for lmrob(y~x), which is robust linear model, "logroblm" for
lmrob(log(y)~x) and "glm", which is generalised linear model, which includes the concept of gamma
distribution to analyse the data.}

\item{result}{The default is "lm". The other option is "glm". If the method are out of "lm", "loglm",
"roblm" and "logroblm", then, result should be defined as "lm". If the method is "glm", then, the
result should be defined as "glm".}
}
\value{
object of Class \code{data.frame}.
}
\description{
Function \code{aptamodel} can be implemented on an object of Class \code{aptakandt},
which is produced as an output from the function \code{CombineReps} and saves the resulting
data to an object of Class \code{data.frame}. See \code{\link[Aptakan]{ReadAptakan}} and
\code{\link[AptaKan]{CombineReps}}. It's an auxilliary function to the function \code{aptamodelall}.
}
\details{
Allows the user to analyse the data on the basis of different models, including linear model,
log linear model, robust linear model, log robust minear model and generalised linear model. It provides
the different values of r-square, adj.rsquare, intercept and slope in case of linear, log linear, robust
linear and log robust linear models as well as the values of AIC, intercept and slope in case of generalised
linear model. It saves the results in slot modelData of an object of Class \code{aptakandt}.
This function \code{aptamodel} is an auxilliary function to the function \code{aptamodelall}.
More details are provided in the\code{AptaKan} package vignette.
}
\examples{
## Here the examples are shown only with the .csv files and combining replicates on the basis of mean.
##Similarly, they can also be implemented on the other results of combining the replicates on the basis of
## median too through function CombineReps, For complete detail, see the vignettes
##To read the .csv file.
file1 <- system.file("exData", "Aptakandata.csv", package = "AptaKan")
read.datacsv <- ReadAptakan(file1)

## To combine the replicates on the basis of mean
cr.mean <- CombineReps(read.datacsv)
cr.mean ## To express the combined replicates

##To analyse the data on the basis of linear model
lm.result <- aptamodel(cr.mean)
lm.result ## To express the result

## To analyse the data on the basis of log linear model
loglm.result <- aptamodel(cr.mean, method = "loglm")
loglm.result ##To express the result

## To analyse the data on the basis of robust linear model
roblm.result <- aptamodel(cr.mean, method = "roblm")
roblm.result ##To express the result

## To analyse the data on the basis of log robust linear model
logroblm.result <- aptamodel(cr.mean, method = "logroblm")
logroblm.result ##To express the result

## To analyse the data on the basis of generalised linear model
glm.result <- aptamodel(cr.mean, method = "glm", result = "glm")
glm.result  ##To express the result
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{CombineReps,}
\keyword{ReadAptaKan,}
\keyword{aptakandt,}
\keyword{aptamodel}
