% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineReps.R
\docType{methods}
\name{CombineReps}
\alias{CombineReps}
\title{To combine the technical replicates}
\usage{
\S4method{CombineReps}{aptakandt}(aptakandt, calc = "Mean")
}
\arguments{
\item{aptakandt}{an object of class \code{"aptakandt"}, which is an output of the function \code{ReadAptakan}.}

\item{calc}{the mathematical method out of "Mean" or "Median" to combine the technical replicates
and "sd" to compute the standard deviation of the replicates. The default method is "Mean".}
}
\value{
object of Class \code{"aptakandt"}, where the result of combined replicates will be saved in the
slot repData.
}
\description{
Function \code{CombineReps} combines the technical replicates (on the basis of mean or
median and their standard deviation) of an object of class \code{"aptakandt"}, which is produced as
an output from the function \code{ReadAptakan} and saves the resulting data to an object of
class \code{"aptakandt"}. See also See \code{\link[Aptakan]{ReadAptakan}} function.
}
\details{
Allows the user to combine technical replicates on the basis of mean and median as well as to
calculate the standard deviation within the replicates. More details are provided in the
\code{AptaKan} package vignette.
}
\examples{
##To read the .csv file
file1 <- system.file("exData", "Aptakandata.csv", package = "AptaKan")
read.datacsv <- ReadAptakan(file1)

## To express all the data
read.datacsv

## To read the .txt file
file1 <- system.file("exData", "Aptakandata.txt", package = "AptaKan")
read.datatxt <- ReadAptakan(file1, type = ".txt")

## To express all the data
read.datatxt

## To combine the replicates on the basis of mean
cr.mean <- CombineReps(read.datacsv)
cr.mean ## To express all the data
slot(cr.mean,"crepData")  ##To express the combined replicates

cr.mean.txt <- CombineReps(read.datatxt)
cr.mean.txt
slot(cr.mean.txt,"crepData")

##To combine the replicates on the basis of meadian
cr.median <- CombineReps(read.datacsv, calc = "Median")
cr.median
slot(cr.median,"crepData")

##To compute the standard deviation within the replicates
cr.sd <- CombineReps(read.datacsv, calc = "sd")
cr.sd ##To express all the data
slot(cr.sd,"crepData") ##To express the standard deviation
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{aptakandt}
