% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kd.gui.R
\name{Kd.gui}
\alias{Kd.gui}
\title{Run the graphical user interface for computing the dissociation constant and
correlation}
\usage{
Kd.gui()
}
\description{
Run the graphical user interface
}
\examples{
Kd.gui()
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
