% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAptakan.R
\name{ReadAptakan}
\alias{ReadAptakan}
\title{To read the .csv as well as .txt file of experimental data}
\usage{
ReadAptakan(file, type = ".csv")
}
\arguments{
\item{file}{the name of the file to read in.}

\item{type}{the type of the file out of ".csv" or ".txt" to read in. The default parameter is ".csv."}
}
\value{
object of Class \code{"aptakandt"} with intialData and concentration slots.
}
\description{
Function \code{ReadAptakan} reads in .csv as well as .txt files of experimental fluorescent
data and uses the data to populate an object of class \code{"aptakandt"}.
}
\details{
Allows the user to read in experimental fluorescence data, arranged in a particular format.
More details are provided in the \code{AptaKan} package vignette.
}
\examples{
##To read the .csv file
file1 <- system.file("exData", "Aptakandata.csv", package = "AptaKan")
read.datacsv <- ReadAptakan(file1)

## To express all the data
read.datacsv

## To express the initial data (all the data read in, except concentration)
slot(read.datacsv,"initialData")

## To express the concentration data (from the data read in)
slot(read.datacsv,"concentration")

## To read the .txt file
file1 <- system.file("exData", "Aptakandata.txt", package = "AptaKan")
read.datatxt <- ReadAptakan(file1, type = ".txt")

## To express the initial data (except concentration)
slot(read.datatxt, "initialData")

## To express the concentration data
slot(read.datatxt, "concentration")
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{ReadAptakan}
\keyword{aptakandt,}
